Add onpush int-* github action
For coin integration branches, it is necessary to trigger this push to
help testers down the road (Live-common CLI tests).
Any push/merge on int-COIN will push a new @ledgerhq/ledger-core@COIN so
that live-common-only users hopefully don't need a full GCC toolchain to
test the current version in CLI.

It does not not run the yarn version on the Action.
yarn version is actually a hidden git commit so it will trigger an
infinite loop (actually it fails because there are no credentials)